+++ src/main/java/org/soomgo/soomgo_project/controller/LoginController.java
@@ -1,8 +1,8 @@
 package org.soomgo.soomgo_project.controller;
 
-import org.soomgo.soomgo_project.domain.UserDTO;
-import org.soomgo.soomgo_project.domain.UserProfileDTO;
-import org.soomgo.soomgo_project.service.UserService;
+import org.soomgo.soomgo_project.domain.userpage.UserDTO;
+import org.soomgo.soomgo_project.domain.userpage.UserProfileDTO;
+import org.soomgo.soomgo_project.service.userpage.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -25,6 +25,7 @@ public class LoginController {
     @PostMapping("/login")
     public String login(@RequestParam String email, @RequestParam String password, HttpSession session, Model model) {
         UserDTO user = userService.getUserByEmailAndPassword(email, password);
+
         if (user != null) {
             session.setAttribute("user", user);
             UserProfileDTO userprofile = userService.getUserProfileByUserNum(user.getUser_num());
@@ -32,7 +33,7 @@ public class LoginController {
             return "redirect:/userpage"; // 로그인 후 사용자 페이지로 리디렉션
         } else {
             model.addAttribute("errorMessage", "이메일 또는 비밀번호가 잘못되었습니다.");
-            return "login"; // 로그인 실패 시 로그인 페이지로 돌아감
+            return "/login"; // 로그인 실패 시 로그인 페이지로 돌아감
         }
     }
     @GetMapping("/login")
