+++ src/main/java/org/soomgo/soomgo_project/controller/userpage/UserPageController.java
@@ -1,9 +1,12 @@
-package org.soomgo.soomgo_project.controller;
+package org.soomgo.soomgo_project.controller.userpage;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.soomgo.soomgo_project.domain.UserDTO;
-import org.soomgo.soomgo_project.domain.UserProfileDTO;
-import org.soomgo.soomgo_project.service.UserService;
+import org.soomgo.soomgo_project.domain.userpage.UserDTO;
+import org.soomgo.soomgo_project.domain.userpage.UserProfileDTO;
+import org.soomgo.soomgo_project.domain.userpage.requestDTO;
+import org.soomgo.soomgo_project.service.userpage.RequestSentService;
+import org.soomgo.soomgo_project.service.userpage.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,18 +16,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpSession;
+import java.util.List;
 
 @RequestMapping("/userpage")
 @Controller
 @Log4j2
+@RequiredArgsConstructor
 public class UserPageController {
 
     private final UserService userService; // UserService 필드 추가
 
-    @Autowired // 생성자 주입을 위한 Autowired 애노테이션 추가
-    public UserPageController(UserService userService) {
-        this.userService = userService;
-    }
+    private final RequestSentService requestSentService;
 
     @GetMapping("")
     public String userPage(HttpSession session, Model model) {
@@ -78,6 +80,17 @@ public class UserPageController {
         return "redirect:/login";
     }
 
-
+    @GetMapping("/requestsent")
+    public String requestSent(HttpSession session, Model model) {
+        UserDTO user = (UserDTO) session.getAttribute("user");
+        if (user != null) {
+            String userEmail = user.getUser_email();
+            List<requestDTO> requests = requestSentService.getRequestsByUserEmail(userEmail);
+            model.addAttribute("requests", requests);
+            return "userpage/requestsent"; // requestsent.jsp로 요청을 포워딩
+        } else {
+            return "redirect:/login"; // 사용자 미로그인 시 로그인 페이지로 리디렉션
+        }
+    }
 
 }
