+++ src/main/java/org/soomgo/soomgo_project/service/userpage/UserService.java
@@ -1,7 +1,7 @@
-package org.soomgo.soomgo_project.service;
+package org.soomgo.soomgo_project.service.userpage;
 
-import org.soomgo.soomgo_project.domain.UserDTO;
-import org.soomgo.soomgo_project.domain.UserProfileDTO;
+import org.soomgo.soomgo_project.domain.userpage.UserDTO;
+import org.soomgo.soomgo_project.domain.userpage.UserProfileDTO;
 
 public interface UserService {
     UserDTO getUserByEmailAndPassword(String email, String password);
