+++ src/main/java/org/soomgo/soomgo_project/service/userpage/UserServiceImpl.java
@@ -1,8 +1,8 @@
-package org.soomgo.soomgo_project.service;
+package org.soomgo.soomgo_project.service.userpage;
 
-import org.soomgo.soomgo_project.domain.UserDTO;
-import org.soomgo.soomgo_project.domain.UserProfileDTO;
-import org.soomgo.soomgo_project.mappers.UserMapper;
+import org.soomgo.soomgo_project.domain.userpage.UserDTO;
+import org.soomgo.soomgo_project.domain.userpage.UserProfileDTO;
+import org.soomgo.soomgo_project.mappers.userpage.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
