+++ src/main/resources/logo.jpeg
