+++ src/main/resources/mappers/userpage/RequestSentMapper.xml
@@ -3,10 +3,10 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.soomgo.soomgo_project.mappers.RequestSentMapper">
+<mapper namespace="org.soomgo.soomgo_project.mappers.userpage.RequestSentMapper">
 
     <!-- 사용자의 이메일로 요청을 조회 -->
-    <select id="getRequestsByUserEmail" parameterType="string" resultType="org.soomgo.soomgo_project.domain.requestDTO">
+    <select id="getRequestsByUserEmail" parameterType="string" resultType="org.soomgo.soomgo_project.domain.userpage.requestDTO">
         SELECT * FROM reqsend
         WHERE writer = #{email}
     </select>
