+++ src/main/resources/mappers/userpage/UserMapper.xml
@@ -3,19 +3,19 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.soomgo.soomgo_project.mappers.UserMapper">
+<mapper namespace="org.soomgo.soomgo_project.mappers.userpage.UserMapper">
 
     <!-- 사용자 이메일과 비밀번호로 사용자 정보 조회 -->
-    <select id="findByEmailAndPassword" parameterType="map" resultType="org.soomgo.soomgo_project.domain.UserDTO">
+    <select id="findByEmailAndPassword" parameterType="map" resultType="org.soomgo.soomgo_project.domain.userpage.UserDTO">
         SELECT * FROM user WHERE user_email = #{email} AND user_password = #{password}
     </select>
 
-    <select id="findUserProfileByUserNum" parameterType="int" resultType="org.soomgo.soomgo_project.domain.UserProfileDTO">
+    <select id="findUserProfileByUserNum" parameterType="int" resultType="org.soomgo.soomgo_project.domain.userpage.UserProfileDTO">
         SELECT * FROM user_profile WHERE user_num = #{user_num}
     </select>
 
 
-    <update id="updateUserProfile" parameterType="org.soomgo.soomgo_project.domain.UserProfileDTO">
+    <update id="updateUserProfile" parameterType="org.soomgo.soomgo_project.domain.userpage.UserProfileDTO">
         UPDATE user_profile
         SET user_nickname = #{user_nickname}
         WHERE user_num = #{user_num}
