--- src/main/webapp/WEB-INF/views/footer.jsp
@@ -1,10 +1,0 @@
-<!-- footer.jsp -->
-<%@ page contentType="text/html; charset=UTF-8" language="java" %>
-<footer>
-    <div class="footer-container">
-        <p>&copy; 숨고카피카피. All rights reserved.</p>
-
-    </div>
-</footer>
-</body>
-</html>
