--- src/main/webapp/WEB-INF/views/header.jsp
@@ -1,17 +1,0 @@
-<!-- header.jsp -->
-<%@ page contentType="text/html; charset=UTF-8" language="java" %>
-<!DOCTYPE html>
-
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="<c:url value='/css/styles.css' />">
-    <title>숨고</title>
-</head>
-<body>
-<header>
-    <div class="header-container">
-        <img src="<c:url value='/WEB-INF/images/logo.jpeg' />" alt="Example Image" />
-    </div>
-</header>
