+++ src/main/webapp/WEB-INF/views/login.jsp
@@ -1,18 +1,16 @@
 <%@ page contentType="text/html; charset=UTF-8" language="java" %>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
-<%@ include file="header.jsp" %>
+<%@ include file="header/header.jsp"%>
 <!DOCTYPE html>
 <html lang="ko">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>로그인</title>
-    <link rel="stylesheet" href="<c:url value='/css/styles.css' />">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/styles.css">
 </head>
 <body>
-<header>
-    <h1>로그인 페이지</h1>
-</header>
+
 <div class="container">
     <main>
         <c:if test="${not empty errorMessage}">
@@ -27,6 +25,6 @@
         </form>
     </main>
 </div>
-<%@ include file="footer.jsp" %>
+
 </body>
 </html>
