+++ src/main/webapp/resources/styles.css
@@ -1,7 +1,7 @@
 /* styles.css */
 
 body {
-    margin: 0;
+    margin : 0;
     padding: 0;
     font-family: Arial, sans-serif;
 }
@@ -70,3 +70,6 @@ footer {
     text-decoration: none;
     cursor: pointer;
 }
+
+/* 드롭다운 */
+
